Update master branch for 1.0 release

+++ app/src/main/java/com/symphony/mrfit/data/model/User.kt
@@ -1,7 +1,7 @@
 /*
- *  Created by Team Symphony on 2/24/23, 11:21 PM
+ *  Created by Team Symphony on 4/24/23, 3:50 AM
  *  Copyright (c) 2023 . All rights reserved.
- *  Last modified 2/24/23, 11:20 PM
+ *  Last modified 4/24/23, 3:49 AM
  */
 
 package com.symphony.mrfit.data.model
@@ -14,7 +14,7 @@ data class User(
     val userID: String = "",
     var name: String? = null,
     var age: Int? = null,
-    var height: Int? = null,
+    var height: Double? = null,
     var weight: Double? = null,
     var visibility: Int = 0, // 0 = Private, 1 = Friends, 2 = Public
     var goals: ArrayList<Goal>? = null,
